% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooks_distance.R
\name{cooks_distance}
\alias{cooks_distance}
\title{Cook's Distance Measure}
\usage{
cooks_distance(model)
}
\arguments{
\item{model}{A linear model object (class lm).}
}
\value{
A vector of Cook's Distance values.
}
\description{
This function calculates Cook's Distance for an lm object.
}
