% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooks_distance.R
\name{custom_hatvalues}
\alias{custom_hatvalues}
\title{Custom Leverage Values (Hat Matrix Diagonal)}
\usage{
custom_hatvalues(model)
}
\arguments{
\item{model}{A linear model object of class 'lm'.}
}
\value{
A numeric vector of leverage values for each observation.
}
\description{
This function calculates the leverage values for each observation in a linear model.
}
