% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooks_distance.R
\name{custom_rstudent}
\alias{custom_rstudent}
\title{Custom Studentized Residuals}
\usage{
custom_rstudent(model)
}
\arguments{
\item{model}{A linear model object of class 'lm'.}
}
\value{
A numeric vector of studentized residuals for each observation.
}
\description{
This function calculates studentized residuals for a linear model object.
}
