% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dffits.R
\name{dffits_measure}
\alias{dffits_measure}
\title{DFFITS Measure}
\usage{
dffits_measure(model)
}
\arguments{
\item{model}{A linear model object (class lm).}
}
\value{
A vector of DFFITS values.
}
\description{
This function calculates DFFITS for an lm object.
}
