% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadis_influence.R
\name{hadis_influence_measure}
\alias{hadis_influence_measure}
\title{Hadi's Influence Measure}
\usage{
hadis_influence_measure(model)
}
\arguments{
\item{model}{A linear model object (class lm).}
}
\value{
A vector of Hadi's Influence Measure values.
}
\description{
This function calculates Hadi's Influence Measure for an lm object.
}
