% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate Inputs for Influence Measure Calculations}
\usage{
validate_inputs(data, model)
}
\arguments{
\item{data}{A data frame or matrix containing the data used in the model.}

\item{model}{A linear model object of class 'lm'.}
}
\value{
NULL. This function is used for its side effect of stopping execution if inputs are invalid.
}
\description{
Validate Inputs for Influence Measure Calculations
}
\examples{
model <- lm(mpg ~ wt + hp, data = mtcars)
validate_inputs(mtcars, model)
}
